Dark Mode: Elevate the theme dropdown from experimental to regular settings
