+++ ui/pages/settings/experimental-tab/experimental-tab.container.js
@@ -6,14 +6,12 @@ import {
   setUseCollectibleDetection,
   setOpenSeaEnabled,
   setEIP1559V2Enabled,
-  setTheme,
 } from '../../../store/actions';
 import {
   getUseTokenDetection,
   getUseCollectibleDetection,
   getOpenSeaEnabled,
   getEIP1559V2Enabled,
-  getTheme,
 } from '../../../selectors';
 import ExperimentalTab from './experimental-tab.component';
 
@@ -25,7 +23,6 @@ const mapStateToProps = (state) => {
     useCollectibleDetection: getUseCollectibleDetection(state),
     openSeaEnabled: getOpenSeaEnabled(state),
     eip1559V2Enabled: getEIP1559V2Enabled(state),
-    theme: getTheme(state),
   };
 };
 
@@ -39,7 +36,6 @@ const mapDispatchToProps = (dispatch) => {
       dispatch(setUseCollectibleDetection(val)),
     setOpenSeaEnabled: (val) => dispatch(setOpenSeaEnabled(val)),
     setEIP1559V2Enabled: (val) => dispatch(setEIP1559V2Enabled(val)),
-    setTheme: (val) => dispatch(setTheme(val)),
   };
 };
